chore: rename classloader modules
