--- fullstack-classloader-api/src/main/java/com/hedera/fullstack/classloader/api/package-info.java
@@ -1,1 +1,0 @@
-package com.hedera.fullstack.classloader.api;
