--- fullstack-classloader-api/src/main/java/module-info.java
@@ -1,1 +1,0 @@
-module com.hedera.fullstack.classloader.api {}
