--- fullstack-classloader-core/src/main/java/com/hedera/fullstack/classloader/core/package-info.java
@@ -1,1 +1,0 @@
-package com.hedera.fullstack.classloader.core;
