--- fullstack-classloader-core/src/test/java/module-info.java
@@ -1,1 +1,0 @@
-module com.hedera.fullstack.classloader.core.test {}
