+++ fullstack-servicelocator-api/build.gradle.kts
