+++ fullstack-servicelocator-api/gradle.properties
