+++ fullstack-servicelocator-api/src/main/java/com/hedera/fullstack/servicelocator/api/package-info.java
@@ -1,0 +1,1 @@
+package com.hedera.fullstack.servicelocator.api;
