+++ fullstack-servicelocator-core/build.gradle.kts
