+++ fullstack-servicelocator-core/gradle.properties
