+++ fullstack-servicelocator-core/src/main/java/module-info.java
@@ -1,0 +1,1 @@
+module com.hedera.fullstack.servicelocator.core {}
